requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.31.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.3
whitenoise==6.3.0
